requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
sqlalchemy
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
motor==3.2.0
pydantic==2.3.0
pymongo==4.4.0
python-dotenv
aiofiles
requests==2.31.0
openai==0.27.8
pillow
boto3==1.28.0
botocore==1.31.0
